feat: restructure project
