+++ components/r3f/ThreeDContent.tsx
@@ -1,18 +1,7 @@
+import { ContactSectionProps } from '@dom/ContactSection'
+import { IntroductionSectionProps } from '@dom/IntroductionSection'
+import { CanvasViews } from '@r3f/CanvasViews'
 import { Canvas } from '@react-three/fiber'
-// import { Scroll, ScrollControls, View } from '@react-three/drei'
-// import { Welcome } from './r3f/Welcome'
-// import { SelfIntroduction } from './r3f/SelfIntroduction'
-// import { ProfessionIntroduction } from './r3f/ProfessionIntroduction'
-// import { Experience } from './r3f/Experience'
-// import { Contact } from './r3f/Contact'
-import {
-	ExperienceCardViews,
-	ExperienceCardViewsProps
-} from './r3f/ExperienceCardViews'
-// import { MediaLinks } from './MediaLinks'
-// import { Acknowledgement } from './Acknowledgement'
-import { IntroductionSectionProps } from './IntroductionSection'
-import { ContactSectionProps } from './ContactSection'
 
 type ThreeDContentProps = {
 	eventSource: React.MutableRefObject<HTMLElement>
@@ -35,11 +24,11 @@ export const ThreeDContent: React.FC<ThreeDContentProps> = ({
 			{/* <gridHelper rotation-x={-Math.PI / 2} /> */}
 
 			{/* Views */}
-			<ExperienceCardViews
+			<CanvasViews
 				experienceTrack={experienceTrack}
 				introductionTrack={introductionTrack}
 				contactTrack={contactTrack}
-			></ExperienceCardViews>
+			></CanvasViews>
 		</Canvas>
 	)
 }
