+++ components/r3f/contact/index.ts
@@ -1,0 +1,1 @@
+export * from './ContactView'
