+++ components/r3f/experience/index.ts
@@ -1,0 +1,2 @@
+export * from './ExperienceView'
+export * from './ThreeJSExperience'
