+++ tsconfig.json
@@ -2,7 +2,13 @@
 	"compilerOptions": {
 		"baseUrl": ".",
 		"paths": {
-			"components": ["components"]
+			"components": ["components"],
+			"@components": ["components"],
+			"@components/*": ["components/*"],
+			"@dom": ["components/dom"],
+			"@dom/*": ["components/dom/*"],
+			"@r3f": ["components/r3f"],
+			"@r3f/*": ["components/r3f/*"]
 		},
 		"target": "es5",
 		"lib": ["dom", "dom.iterable", "esnext"],
